Further adventures in sdk coding (#3648)
* Updates CHANGES file

* Prepares the way for capturing Blknum

* Fixes build

* Further adventures

* Further adventures

* Further adventures

* Further adventures

* Further adventures

* Further adventures

* Trying to capture MustParseInt

* Captures MustParse...

* Changes status size, makes Blknum a Numeral

* Commits to Numeral block numbers

* Fixes build

* Catches up with auto-code gen

* Commit to concrete type of Numeral

* Better docs, remove double in favor of float64 from auto-code-gen

* Validate codebase before generating code

+++ src/apps/chifra/internal/traces/handle_counts.go
@@ -88,7 +88,7 @@ func (opts *TracesOptions) HandleCounts() error {
 
 				for _, item := range items {
 					counter := types.TraceCount{
-						BlockNumber:      uint64(item.BlockNumber),
+						BlockNumber:      item.BlockNumber,
 						TransactionIndex: item.TransactionIndex,
 						TransactionHash:  item.Hash,
 						Timestamp:        item.Timestamp,
